Start by separating the code into a module:
This commit just moves code around. Nothing to worry about.

All the code is moved to lib as a module with a method ("run"),
and the version variable normalized to $VERSION (the common
practice, so ->VERSION method from UNIVERSAL works), and then the
command line uses the module.

Additional steps would include:
* Proper separation of subroutines.
* Separating the getopt, moving it to either a class or the CLI
* Separating the external applications, seeing what could be
  done in Perl instead to improve portability
* Create proper entities and components and resolve inner
  dependencies