Take all system commands variables outside run() method:
This commits takes the variables generated for all the system commands
needed and puts them outside the "run" method.

Instead of creating multiple variables, we create a hash with all
of them as entries using "map". We replace all uses of the variables
to the new hash.

This shows a few possible issues (some variables aren't used, others
get overwritten - such as $file, some variables exist but the binary
is still used directly - such as with "rm").